requirements.txt
Flask==3.1.2
tensorflow==2.10.0
Pillow==10.0.0
numpy<2.0
Werkzeug==3.1.3
Jinja2==3.1.6
click==8.2.1
itsdangerous==2.2.0
blinker==1.9.0
colorama==0.4.6
MarkupSafe==3.0.2
